+++ lib/foundation/constants/routes.dart
@@ -1,0 +1,3 @@
+class ToDoRoutes {
+  static String todoHomeScreen = 'todoHomeScreen';
+}
